requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.12

# Environment & Config
python-dotenv==1.0.1

# Web Scraping & Data Collection
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
PyPDF2==3.0.1

# AI & LLM
anthropic==0.39.0
openai==1.54.0

# AWS Services
boto3==1.35.0
botocore==1.35.0
mangum==0.18.0

# Vector Database & RAG
chromadb==0.5.18
sentence-transformers==3.2.1

# Database
psycopg2-binary
sqlalchemy==2.0.36

# SMS/Communication
twilio

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# HTTP Client
httpx==0.27.2

# Testing
pytest==8.3.3
